Move SpreadsheetExportMixin to generic.IndexView
