+++ wagtail/admin/templates/wagtailadmin/generic/index.html
@@ -1,8 +1,13 @@
 {% extends "wagtailadmin/generic/listing.html" %}
-{% load i18n %}
+{% load i18n wagtailadmin_tags %}
 
 {% block header %}
-    {% include "wagtailadmin/shared/header.html" with title=page_title subtitle=page_subtitle action_url=header_action_url action_text=header_action_label icon=header_icon search_url=search_url search_form=search_form search_results_url=index_results_url only %}
+    {% fragment as extra_actions %}
+        {% if view.list_export %}
+            {% include view.export_buttons_template_name %}
+        {% endif %}
+    {% endfragment %}
+    {% include "wagtailadmin/shared/header.html" with title=page_title subtitle=page_subtitle action_url=header_action_url action_text=header_action_label extra_actions=extra_actions icon=header_icon search_url=search_url search_form=search_form search_results_url=index_results_url only %}
 {% endblock %}
 
 {% block listing %}
