+++ wagtail/admin/views/reports/base.py
@@ -1,12 +1,9 @@
 from django.utils.translation import gettext_lazy as _
 
 from wagtail.admin.views.generic.models import IndexView
-from wagtail.admin.views.mixins import SpreadsheetExportMixin
 
 
-class ReportView(SpreadsheetExportMixin, IndexView):
-    header_icon = ""
-    page_kwarg = "p"
+class ReportView(IndexView):
     template_name = "wagtailadmin/reports/base_report.html"
     title = ""
     paginate_by = 50
@@ -20,23 +17,13 @@ class ReportView(SpreadsheetExportMixin, IndexView):
 
     def get(self, request, *args, **kwargs):
         self.filters, self.object_list = self.get_filtered_queryset()
-        if self.is_export:
-            self.object_list = self.decorate_paginated_queryset(self.object_list)
-            return self.as_spreadsheet(self.object_list, self.request.GET.get("export"))
-        else:
-            context = self.get_context_data()
-            context["object_list"] = self.decorate_paginated_queryset(
-                context["object_list"]
-            )
-            return self.render_to_response(context)
+        self.object_list = self.decorate_paginated_queryset(self.object_list)
+        context = self.get_context_data()
+        return self.render_to_response(context)
 
-    def get_context_data(self, *args, object_list=None, **kwargs):
-        queryset = object_list if object_list is not None else self.object_list
-
-        context = super().get_context_data(*args, object_list=queryset, **kwargs)
+    def get_context_data(self, *args, **kwargs):
+        context = super().get_context_data(*args, **kwargs)
         context["title"] = self.title
-        context["header_icon"] = self.header_icon
-        context["filters"] = self.filters
         return context
